external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/divide_int_mat2.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:4: '/' :  wrong operand types  no operation '/' exists that takes a left-hand operand of type 'const mediump 2X2 matrix of float' and a right operand of type 'const mediump int' (or there is no acceptable conversion)

implicit cast of mat2 divided by int should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast of mat2 divided by int should fail
test vertex shader
default fragment shaderPASS implicit cast of mat2 divided by int should fail
PASS successfullyParsed is true

TEST COMPLETE
